add save_model to Accelerator object
